+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add support for barrier holes on "Start Reversed" offset glass/grating.
 * Added some additional compatibility logic to make styles with old corridor definitions work in some additional cases.
 * Switch more package loading errors to be displayed in the more friendly error window.
+* Add support for Connection Signage producing blue/orange frames.
 
 # Version 4.46.1
 * Fix FGD export producing invalid syntax.